requirements.txt
numpy
scipy
pandas
pillow
matplotlib
scikit-image